Move FaceRec class to separate file, adjust paths accordingly
