+++ tests/test_facerec.py
@@ -7,7 +7,7 @@ import os
 from PIL import Image, ImageChops, ImageStat
 from io import BytesIO
 import face_recognition 
-from reproduce.test import FaceRec
+from reproduce.face_rec import FaceRec
 
 face_image_link = "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=687&q=80"
 failure_mode_image_link= "https://www.cdc.gov/healthypets/images/pets/cute-dog-headshot.jpg"
